+++ test/JsonApiDotNetCoreTests/IntegrationTests/EagerLoading/Street.cs
@@ -7,7 +7,7 @@ namespace JsonApiDotNetCoreTests.IntegrationTests.EagerLoading;
 
 [UsedImplicitly(ImplicitUseTargetFlags.Members)]
 [Resource(ControllerNamespace = "JsonApiDotNetCoreTests.IntegrationTests.EagerLoading")]
-public sealed class Street : Identifiable<int>
+public sealed class Street : Identifiable<long>
 {
     [Attr]
     public string Name { get; set; } = null!;
