+++ test/JsonApiDotNetCoreTests/IntegrationTests/ExceptionHandling/ThrowingArticle.cs
@@ -7,7 +7,7 @@ namespace JsonApiDotNetCoreTests.IntegrationTests.ExceptionHandling;
 
 [UsedImplicitly(ImplicitUseTargetFlags.Members)]
 [Resource(ControllerNamespace = "JsonApiDotNetCoreTests.IntegrationTests.ExceptionHandling")]
-public sealed class ThrowingArticle : Identifiable<int>
+public sealed class ThrowingArticle : Identifiable<long>
 {
     [Attr]
     [NotMapped]
